requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.6
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
cryptography==2.4.2
idna==2.8
paramiko==2.4.2
pyasn1==0.4.5
pycparser==2.19
PyNaCl==1.3.0
PyYAML==5.1
requests==2.21.0
six==1.12.0
urllib3==1.24.1
wincertstore==0.2
